Gate token revocation behind the Rest module
I consider this to be REST functionality, and so to reduce bundle size
it shouldn’t be bundled with a BaseRealtime client. I should have done
this in 89c0761.
